miApiJWT/requirements.txt
uvicorn[standard]
fastapi
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1